move progressive item UI data into game folders
