+++ randovania/games/dread/item_database/progressive_items.py
@@ -1,0 +1,27 @@
+def gui_tuples():
+    return [
+        (
+            "Progressive Beam",
+            ("Wide Beam", "Plasma Beam", "Wave Beam")
+        ),
+        (
+            "Progressive Charge Beam",
+            ("Charge Beam", "Diffusion Beam")
+        ),
+        (
+            "Progressive Missiles",
+            ("Super Missiles", "Ice Missiles")
+        ),
+        (
+            "Progressive Suit",
+            ("Varia Suit", "Gravity Suit")
+        ),
+        (
+            "Progressive Bomb",
+            ("Bomb", "Cross Bomb")
+        ),
+        (
+            "Progressive Spin",
+            ("Spin Boost", "Space Jump")
+        )
+    ]
