+++ randovania/games/prime2/item_database/prime2_progressive_items.py
@@ -1,0 +1,11 @@
+def gui_tuples():
+    return [
+        (
+            "Progressive Suit",
+            ("Dark Suit", "Light Suit")
+        ),
+        (
+            "Progressive Grapple",
+            ("Grapple Beam", "Screw Attack")
+        )
+    ]
