+++ randovania/games/prime3/item_database/prime3_progressive_items.py
@@ -1,0 +1,11 @@
+def gui_tuples():
+    return [
+        (
+            "Progressive Missile",
+            ("Ice Missile", "Seeker Missile")
+        ),
+        (
+            "Progressive Beam",
+            ("Plasma Beam", "Nova Beam")
+        )
+    ]
